* Added a bid function to let workers bid for a job. * Admins can mark a job as completed and workers cannot bid or do completed jobs. * Add a page for editors/reviewers to review survey answers and choose which video to crawl into local.

+++ pyano2/class_views/vatic.py
@@ -2,10 +2,15 @@ from django.conf import settings
 from django.http import JsonResponse
 from django.shortcuts import render, redirect
 from django.views import View
+from django.core.mail import send_mail
+from django.db.models import Count
 
 from pyano2.models import *
-
+from pyano2.downloader.youtube import download_youtube_video
+from pyano2.downloader.video import *
+from pyano import settings as pyano_settings
 import json
+from glob import glob
 
 
 class VATICIndexView(View):
@@ -30,13 +35,14 @@ class VATICIndexView(View):
                 return redirect(to="/vatic/list")
         return render(request, template_name=self.template_name, context=context)
 
+
 class VATICListView(View):
     template_name = 'vatic/list.html'
 
     def get(self, request, *args, **kwargs):
         if not request.user.is_authenticated:
             return redirect(to="{}?next=/vatic/list".format(settings.LOGIN_URL))
-        jobs = VATICJob.objects.all()
+        jobs = VATICJob.objects.filter(completed=False).all()
         return render(request, template_name=self.template_name, context={'jobs': jobs})
 
 
@@ -166,6 +172,11 @@ class VATICSaveJobView(View):  # savejob
         jobs = VATICJob.objects.filter(id=id)
         if len(jobs) > 0:
             job = jobs[0]
+            assignments = VATICWorkerJob.objects.filter(worker=request.user, job=job)
+            if assignments.count() < 1:
+                return JsonResponse({'error': 'User is not assigned for this job.'})
+            if job.completed:
+                return JsonResponse({'error': 'Job is completed. You cannot submit after a job is finalized.'})
             for path in job.paths.all():
                 path.delete()
             self.readpaths(job, tracks)
@@ -231,3 +242,222 @@ class VATICValidateJobView(View):  # validatejob: to validate whether if annotat
             job = jobs[0]
             matched = job.validator(paths, job.paths)
         return JsonResponse({'status': matched})
+
+
+class VATICBidJobView(View):
+    template_name = 'vatic/bid.html'
+
+    def get(self, request, *args, **kwargs):
+        if not request.user.is_authenticated:
+            return redirect(to="{}?next=/vatic/bid?id={}".format(settings.LOGIN_URL, request.GET.get('id')))
+        context = {}
+        try:
+            id = request.GET.get('id')
+            wjobs = request.user.worker2jobs.all()
+            assigned = False
+            for wjob in wjobs:
+                if wjob.job.id == int(id):
+                    assigned = True
+            bids = VATICBid.objects.filter(job_id=id, candidate=request.user)
+            if bids.count() > 0:
+                if not assigned:
+                    if not request.user.is_staff:
+                        context['error'] = 'You already bid this job. Probably admins are still reviewing your application.'
+            elif not assigned:
+                job = VATICJob.objects.filter(id=id)[0]
+                num_workers = job.job2workers.count()
+                if num_workers > 0:
+                    context['error'] = 'This position is filled.'
+                elif job.completed:
+                    context['error'] = 'This job is completed. You cannot bid.'
+                else:
+                    bid = VATICBid()
+                    bid.candidate = request.user
+                    bid.job = job
+                    bid.approved = False
+                    bid.approved_by = None
+                    bid.save()
+        except Exception as e:
+            context['error'] = 'Internal Server Error: {}'.format(e)
+        context['assigned'] = assigned
+        return render(request, template_name=self.template_name, context=context)
+
+
+class VATICListJobApplicationView(View):
+    template_name = 'vatic/job_applications.html'
+
+    def get(self, request, *args, **kwargs):
+        if not request.user.is_authenticated:
+            return redirect(to="{}?next=/vatic/job_applications/".format(settings.LOGIN_URL))
+        if not request.user.is_staff:
+            return redirect(to="/")
+        context = {}
+        try:
+            bids = VATICBid.objects.filter(approved=False)
+            context['bids'] = bids
+        except Exception as e:
+            context['error'] = 'Internal Server Error: {}'.format(e)
+        return render(request, template_name=self.template_name, context=context)
+
+
+class VATICApproveBidView(View):
+    template_name = 'vatic/approve_bid.html'
+
+    def email(self, bid, sender):
+        subject = 'Your bid for job ID {} in Shoplifting prevention project is approved'.format(bid.job.id)
+        message = 'Dear {},\n\n' \
+                  'I am {} {} from Shoplifting prevention project. ' \
+                  'Our admins have approved your bid for the job /vatic?id={}. ' \
+                  'You can start to work by following this URL: http://13.58.121.50:8000/vatic/?id={} . \n\n' \
+                  'Thank you very much,\n' \
+                  '{} {}'.format(bid.candidate.username, sender.first_name, sender.last_name,
+                                 bid.job.id, bid.job.id,
+                                 sender.first_name, sender.last_name)
+        email_from = settings.EMAIL_HOST_USER
+        recipient_list = [bid.candidate.email]
+        send_mail(subject, message, email_from, recipient_list)
+        return redirect('/')
+
+    def get(self, request, *args, **kwargs):
+        if not request.user.is_authenticated:
+            return redirect(to="{}?next=/vatic/approve_bid/?id=".format(settings.LOGIN_URL, request.GET.get('id')))
+        if not request.user.is_staff:
+            return redirect(to="/")
+        context = {}
+        try:
+            bids = VATICBid.objects.filter(id=request.GET.get('id'))
+            for bid in bids.all():
+                if bid.approved: continue
+                assignments = VATICWorkerJob.objects.filter(worker=bid.candidate, job=bid.job)
+                if len(assignments) > 0 or bid.job.completed:
+                    context['error'] = 'This job has been assigned to another user or is already completed. ' \
+                                       'This bid cannot be approved!'
+                else:
+                    bid.approved = True
+                    bid.approved_by = request.user
+                    assignment = VATICWorkerJob()
+                    assignment.worker = bid.candidate
+                    assignment.job = bid.job
+                    bid.save()
+                    assignment.save()
+                    self.email(bid, request.user) # send email to the candidate
+                    # remove other applications for the same job
+                    other_bids = VATICBid.objects.filter(job=bid.job)
+                    for o in other_bids.all():
+                        if o.candidate != bid.candidate:
+                            o.delete()
+            if bids.count() == 0:
+                context['error'] = 'No bid found.'
+            else:
+                context['bid'] = bids[0]
+        except Exception as e:
+            context['error'] = 'Internal Server Error: {}'.format(e)
+        return render(request, template_name=self.template_name, context=context)
+
+
+class VATICFinalizeJobView(View):
+    def get(self, request, *args, **kwargs):
+        if not request.user.is_authenticated:
+            return redirect(to="{}?next=/vatic/finalize/".format(settings.LOGIN_URL))
+        if not request.user.is_staff:
+            return redirect(to="/")
+        context = {}
+        id = request.GET.get('id')
+        try:
+            jobs = VATICJob.objects.filter(id=id)
+            for job in jobs.all():
+                if job.completed: continue
+                job.completed = True
+                job.save()
+                break
+        except Exception as e:
+            logging.error('Internal Server Error: {}'.format(e))
+        return redirect(to="/vatic/list")
+
+
+class VATICCrawlerView(View):
+    template_name = 'vatic/crawler.html'
+
+    def get(self, request, *args, **kwargs):
+        if not request.user.is_authenticated:
+            return redirect(to="{}?next=/vatic/crawler/".format(settings.LOGIN_URL))
+        if not request.user.is_staff:
+            return redirect(to="/")
+        context = {}
+        try:
+            gid = request.GET.get('gid')
+            context['gid'] = gid
+            groups = VATICJobGroup.objects.filter(id=gid)
+            if len(groups) > 0:
+                group = groups[0]
+                context['group'] = group
+                videos = Video.objects.annotate(num_responses=Count('responses'))
+                videos = videos.filter(num_responses__gt=0)
+                context['videos'] = videos
+            else:
+                context['error'] = 'No group found.'
+        except Exception as e:
+            context['error'] = 'Internal Server Error: {}'.format(e)
+        return render(request, template_name=self.template_name, context=context)
+
+    def post(self, request, *args, **kwargs):
+        if not request.user.is_authenticated:
+            return JsonResponse({'status': 404, 'error': 'No user'})
+        if not request.user.is_staff:
+            return JsonResponse({'status': 404, 'error': 'No staff'})
+        try:
+            id = request.POST.get('id')
+            vid = request.POST.get('vid')
+            gid = request.POST.get('gid')
+            groups = VATICJobGroup.objects.filter(id=gid)
+            if len(groups) == 0:
+                return JsonResponse({'status': 404, 'error': 'No group'})
+            else:
+                group = groups[0]
+            download_youtube_video(youtube_ids=[vid],
+                                   output_folder=os.path.join(pyano_settings.BASE_DIR, 'static/videos'))
+            files = glob(os.path.join(pyano_settings.BASE_DIR, 'static/videos/{}'.format(request.POST.get('vid'))) + '.*')
+            if len(files) > 0:
+                f = files[0]
+                output_path = os.path.join(pyano_settings.BASE_DIR, 'static/frames/{}'.format(request.POST.get('vid')))
+                extractor = VATICExtractor(video_path=f, output_path=output_path)
+                extractor()
+                labels = request.POST.get('labels').split(',')
+                print(labels)
+                loader = VATICLoader(location=output_path,
+                                     labels=labels,
+                                     pyano_video_id=id)
+                loader(group)
+        except Exception as e:
+            return JsonResponse({'status': 404, 'error': e})
+        return JsonResponse({'status': 200})
+
+
+class VideoAnswerView(View):
+    def post(self, request, vid, *args, **kwargs):
+        if not request.user.is_authenticated:
+            return JsonResponse({'status': 404, 'error': 'No user'})
+        if not request.user.is_staff:
+            return JsonResponse({'status': 404, 'error': 'No staff'})
+        answers = {}
+        try:
+            videos = Video.objects.filter(id=vid)
+            if len(videos) > 0:
+                video = videos[0]
+                responses = video.responses.all()
+                for i, response in enumerate(responses):
+                    answers[i] = {'uname': '', 'questions': []}
+                    if not response.user:
+                        answers[i]['uname'] = 'Unknown'
+                    else:
+                        answers[i]['uname'] = response.user.username
+                    ans = response.answers.all()
+                    for an in ans:
+                        q = an.question.text
+                        a = an.body
+                        answers[i]['questions'].append({'question': q, 'answer': a})
+        except Exception as e:
+            return JsonResponse({'status': 404, 'error': e})
+        return JsonResponse({'status': 200, 'answers': answers})
+
+
